refactor(coach): split CoachDriver into a scheduler and an attempt runner
Closes #208. Phase 5 of wiki/lean-coaching-core.md.

The 1,333-line facade becomes two owners with a stated state-ownership
boundary.

`CoachDriver` is the scheduler and stays the session's public face: it
owns trigger coalescing and pending-trigger generations, transcription
settlement, the single-flight handling slot, and forward-only route
state — selection, advance, skip, exhaustion, and the delivery tokens
that make a terminal transition land exactly once, all under its one
lock.

`CoachAttemptRunner` executes one attempt against one snapshotted
target: filler classification, the bounded tool loop, the capture_screen
continuation, overlay delivery, history commit, off-path compaction with
its own cancellation lifecycle, and attempt identity. Its lock guards
state that never touches scheduling state.

`CoachTranscriptLedger` holds the one datum both halves share — the
committed transcript boundary. It only grows, so it is a leaf lock and
neither half ever holds the other's. The interface is one call per
attempt: an immutable AttemptBrain plus pending work in, an
AttemptExecution out. No new lifecycle state, retries, timers, or
cross-component coordination.

Structure changed; observable coaching did not. Not a single test file
changed in this slice — every existing CoachDriver test passes
unmodified, which is the strongest available proof the split is
behavior-preserving.

Gate: swift build && ./scripts/run-tests.sh — 814 tests in 96 suites passed.
