+++ Sources/JarvisCore/Coach/CoachDriver.swift
@@ -1,31 +1,42 @@
 import Foundation
 
-/// The single-flight coaching coordinator.
+/// The single-flight coaching scheduler, and the session's public face.
 ///
 /// A coaching attempt snapshots exactly one route target and owns that target for its complete tool
 /// loop. Failed provider requests are never replayed inside the attempt. Instead, failed conversation
-/// work stays uncommitted and this coordinator schedules a fresh attempt from the latest finalized
+/// work stays uncommitted and this scheduler starts a fresh attempt from the latest finalized
 /// transcript plus provider-neutral observations completed earlier.
 ///
-/// `@unchecked Sendable` is justified because mutable coordinator state is guarded by `stateLock`;
-/// `CoachHistory`, transcript, and injected adapters are independently synchronized.
+/// The state-ownership boundary with `CoachAttemptRunner`, which executes those attempts
+/// (wiki/lean-coaching-core.md, Phase 5):
+///
+/// - **This type owns** trigger coalescing and pending-trigger generations, transcription
+///   settlement, the single-flight handling slot, and forward-only route state — selection,
+///   advance, skip, exhaustion, and the delivery tokens that make a terminal transition land
+///   exactly once. All of it lives under this type's one `stateLock`.
+/// - **The runner owns** one attempt's execution: filler classification, the bounded tool loop, the
+///   `capture_screen` continuation, overlay delivery, history commit, off-path compaction, and
+///   attempt identity.
+///
+/// The interface is one call per attempt — an immutable `AttemptBrain` plus the pending work in, an
+/// `AttemptExecution` out — over the shared `CoachTranscriptLedger`, which is the single datum both
+/// halves read: the committed transcript boundary that makes a late turn-end idempotent. Nothing
+/// new coordinates between them.
+///
+/// `@unchecked Sendable` is justified because mutable scheduler state is guarded by `stateLock`;
+/// the ledger, the runner, and injected adapters are independently synchronized.
 public final class CoachDriver: @unchecked Sendable {
     public typealias AutomaticAttemptDelay =
         @Sendable (_ consecutiveAutomaticAttempt: Int) async throws -> Void
 
-    private let config: Config
-    private let transcript: RollingTranscript
-    private let screen: ScreenCapturing
-    private let overlay: OverlayRendering
-    private let clock: Clock
-    private let sessionStart: TimeInterval
-    private let history = CoachHistory()
     private let transcriptionSettlement = TranscriptionSettlementGate()
     private let automaticAttemptDelay: AutomaticAttemptDelay
     private let coachingAttempts: (any CoachingAttemptAuditing)?
-
-    /// Safety backstop against a pathological model that loops on capture_screen forever.
-    private let maxToolIterations = 4
+    /// The committed transcript boundary shared with the runner. It only grows, so reading it
+    /// inside this type's lock needs no coordination with the runner's.
+    private let ledger = CoachTranscriptLedger()
+    /// The attempt half. One per driver, so history and the compaction lifetime match the session's.
+    private let runner: CoachAttemptRunner
 
     private let stateLock = NSLock()
     /// The control-plane snapshot new attempts run against. Installed at Start and at explicit
@@ -47,23 +58,12 @@ public final class CoachDriver: @unchecked Sendable {
     private var exhaustionDeliveryGeneration: UInt = 0
     private var pendingExhaustionDeliveryGeneration: UInt?
     private var isHandling = false
-    /// Guards the single off-path compaction run (see `startCompactionIfIdle`).
-    private var isCompacting = false
-    /// A compaction asked for while one was already running, run once the current pass ends.
-    private var compactionRequested = false
-    /// The in-flight pass, so session teardown can cancel and drain it.
-    private var compactionTask: Task<Void, Never>?
-    /// Latched by teardown: no further background pass may start on a stopped session.
-    private var backgroundWorkStopped = false
-    /// Transcript is committed only by a complete, non-truncated `speak` or `stay_silent`.
-    private var committedTranscriptCount = 0
     /// Natural triggers coalesce while an attempt or automatic pending-work wait owns the slot.
     private var pendingTrigger: PendingTrigger?
     /// Monotonic pulse used to race bounded automatic backoff against a newly coalesced trigger
     /// without losing a trigger that lands just before the async waiter is installed.
     private var pendingTriggerGeneration: UInt = 0
     private var pendingTriggerWaiters: [UUID: CheckedContinuation<Void, Never>] = [:]
-    private var nextAttemptID = 0
 
     /// A turn-end carries the transcript boundary that caused it. This makes a delayed callback
     /// idempotent after another admitted attempt already committed the same finalized speech.
@@ -87,39 +87,8 @@ public final class CoachDriver: @unchecked Sendable {
         let onSelected: (@MainActor @Sendable (BrainTarget) -> Void)?
     }
 
-    private struct PendingCoachingWork {
-        var reason: TriggerReason
-        /// Only an explicit hotkey wake may cross unsettled transcription. A retry of a failed
-        /// manual hint is automatic and resets this bit unless another hotkey press joins it.
-        var bypassesTranscriptionSettlement: Bool
-        var wake: CoachingAttemptAuditEvent.Wake = .trigger
-        /// Completed effects safe to carry between attempts and providers: ordinary user context
-        /// only. At most the latest screen observation is retained. Never raw reasoning, tool ids,
-        /// or call/result linkage.
-        var observations: [ChatMessage] = []
-        var manualHintPrepared = false
 
-        init(reason: TriggerReason) {
-            self.reason = reason
-            bypassesTranscriptionSettlement = reason == .manualHint
-        }
-    }
 
-    private enum AttemptResult {
-        case completed(TurnOutcome)
-        case failed(
-            outcome: TurnOutcome,
-            failure: BrainFailure,
-            work: PendingCoachingWork
-        )
-        case skipped(TurnOutcome)
-        case cancelled
-    }
-
-    private struct AttemptExecution {
-        let id: Int?
-        let result: AttemptResult
-    }
 
     // Module-visible only because `BrainSelectionStep` carries one; see the note below.
     struct RouteExhaustionDelivery {
@@ -186,16 +155,20 @@ public final class CoachDriver: @unchecked Sendable {
     ) {
         self.plan = plan
         self.activity = activity
-        self.config = config
-        self.transcript = transcript
         self.configuredRoute = route
         self.routeSession = BrainRouteSession(targetCount: route.targets.count)
-        self.screen = screen
-        self.overlay = overlay
-        self.clock = clock
         self.coachingAttempts = coachingAttempts
-        self.sessionStart = sessionStart ?? clock.now()
         self.automaticAttemptDelay = automaticAttemptDelay ?? Self.defaultAutomaticAttemptDelay
+        self.runner = CoachAttemptRunner(
+            config: config,
+            transcript: transcript,
+            screen: screen,
+            overlay: overlay,
+            clock: clock,
+            sessionStart: sessionStart ?? clock.now(),
+            coachingAttempts: coachingAttempts,
+            activity: activity,
+            ledger: ledger)
     }
 
     private static let defaultAutomaticAttemptDelay: AutomaticAttemptDelay = { sequence in
@@ -322,6 +295,17 @@ public final class CoachDriver: @unchecked Sendable {
         return true
     }
 
+    /// Cancel background work that must not outlive the session, handing back the task so teardown
+    /// can drain it before sealing the audit.
+    ///
+    /// Compaction runs off the attempt path, so `TurnTaskBox` does not own it. Without this a
+    /// summary keeps a provider process alive and billing after Stop, and can still be writing when
+    /// the session audit closes. The runner owns that lifecycle; this is the session's handle on it.
+    @discardableResult
+    public func cancelBackgroundWork() -> Task<Void, Never>? {
+        runner.cancelBackgroundWork()
+    }
+
     /// Provider state feeds both speakers into one aggregate gate. `hasPendingWork` means the provider
     /// still owns speech or transcript work, not merely that the audio waveform is non-silent.
     public func updateTranscriptionWork(_ hasPendingWork: Bool, for speaker: Speaker) {
@@ -332,8 +316,8 @@ public final class CoachDriver: @unchecked Sendable {
     /// reaches this boundary: the first trigger, a trigger queued behind an in-flight model call,
     /// and a pending-work retry. A manual hint is the explicit immediate exception.
     private func waitForTranscriptionSettlement(
-        before work: PendingCoachingWork
-    ) async -> PendingCoachingWork {
+        before work: CoachAttemptRunner.PendingCoachingWork
+    ) async -> CoachAttemptRunner.PendingCoachingWork {
         guard !work.bypassesTranscriptionSettlement else { return work }
 
         let interruptGeneration = transcriptionSettlement.interruptGenerationSnapshot()
@@ -362,28 +346,6 @@ public final class CoachDriver: @unchecked Sendable {
             settledWork.wake = .trigger
         }
         return settledWork
-    }
-
-    private func currentCommittedTranscriptCount() -> Int {
-        stateLock.lock()
-        defer { stateLock.unlock() }
-        return committedTranscriptCount
-    }
-
-    private func commitTranscript(through count: Int) {
-        stateLock.lock()
-        if count > committedTranscriptCount {
-            committedTranscriptCount = count
-        }
-        stateLock.unlock()
-    }
-
-    private func takeNextAttemptID() -> Int {
-        stateLock.lock()
-        nextAttemptID += 1
-        let id = nextAttemptID
-        stateLock.unlock()
-        return id
     }
 
     private enum TriggerClaim {
@@ -456,10 +418,10 @@ public final class CoachDriver: @unchecked Sendable {
         return PendingTrigger(reason: reason, transcriptBoundary: boundary)
     }
 
-    /// Must be called while `stateLock` is held.
+    /// Must be called while `stateLock` is held. The ledger is a leaf: it never takes this lock.
     private func isCoveredByCommittedTranscript(_ trigger: PendingTrigger) -> Bool {
         trigger.reason == .turnEnd
-            && trigger.transcriptBoundary.map { $0 <= committedTranscriptCount } == true
+            && trigger.transcriptBoundary.map { $0 <= ledger.committedCount } == true
     }
 
     /// Atomically take a coalesced trigger and the pulse generation at the same boundary. A later
@@ -803,7 +765,7 @@ public final class CoachDriver: @unchecked Sendable {
             return .brainError
         }
 
-        var work = PendingCoachingWork(reason: reason)
+        var work = CoachAttemptRunner.PendingCoachingWork(reason: reason)
         var automaticSequence = 0
         var latestOutcome: TurnOutcome = .silentByModel
 
@@ -818,7 +780,7 @@ public final class CoachDriver: @unchecked Sendable {
                 return .brainError
             }
 
-            let execution = await runAttempt(work, using: attempt)
+            let execution = await runner.runAttempt(work, using: attempt)
             switch execution.result {
             case .completed(let outcome):
                 latestOutcome = outcome
@@ -834,7 +796,7 @@ public final class CoachDriver: @unchecked Sendable {
                 guard let next = finishOrTakeNextTrigger() else {
                     return latestOutcome
                 }
-                work = PendingCoachingWork(reason: next.reason)
+                work = CoachAttemptRunner.PendingCoachingWork(reason: next.reason)
 
             case .skipped(let outcome):
                 latestOutcome = outcome
@@ -845,7 +807,7 @@ public final class CoachDriver: @unchecked Sendable {
                 guard let next = finishOrTakeNextTrigger() else {
                     return latestOutcome
                 }
-                work = PendingCoachingWork(reason: next.reason)
+                work = CoachAttemptRunner.PendingCoachingWork(reason: next.reason)
 
             case .cancelled:
                 if let id = execution.id {
@@ -928,418 +890,6 @@ public final class CoachDriver: @unchecked Sendable {
 
         releaseHandlingSlot()
         return .cancelled
-    }
-
-    /// Run one attempt on one immutable target snapshot.
-    private func runAttempt(
-        _ pendingWork: PendingCoachingWork,
-        using attempt: AttemptBrain
-    ) async -> AttemptExecution {
-        if Task.isCancelled {
-            jlog("… attempt cancelled (stopped) before handling")
-            return AttemptExecution(id: nil, result: .cancelled)
-        }
-
-        var work = pendingWork
-        let reason = work.reason
-        if case .silence(let seconds) = reason {
-            jlog("🤫 quiet for \(Int(seconds))s")
-        }
-
-        let now = clock.now()
-        let context = TriggerContext(
-            reason: reason,
-            sessionElapsedSeconds: now - sessionStart)
-        let transcriptStartIndex = currentCommittedTranscriptCount()
-        let delta = transcript.renderFrom(index: transcriptStartIndex)
-        let classifications = delta.lines.map { TurnSubstance.classification(of: $0.text) }
-        let brainFacingOffsets = classifications.indices.filter {
-            classifications[$0].isSubstantive
-        }
-        let substantiveLines = brainFacingOffsets.map { delta.lines[$0] }
-        let brainFacingTranscriptIndices = Set(brainFacingOffsets.map {
-            transcriptStartIndex + $0
-        })
-        let substantiveDeltaText = RollingTranscript.render(substantiveLines)
-        let attemptID = takeNextAttemptID()
-        coachingAttempts?.recordStarted(
-            attemptID: attemptID,
-            wake: work.wake,
-            reason: reason,
-            target: attempt.target,
-            transcriptStartIndex: transcriptStartIndex,
-            transcriptLines: delta.lines,
-            classifications: classifications,
-            brainFacingTranscriptIndices: brainFacingTranscriptIndices)
-
-        let userText = [
-            substantiveDeltaText.isEmpty
-                ? nil
-                : JarvisPrompts.Coach.newSpeech(substantiveDeltaText),
-            context.promptLine,
-        ].compactMap { $0 }.joined(separator: "\n\n")
-        var turnMessages = userText.isEmpty ? work.observations : [.user(userText)] + work.observations
-
-        // A request needs meaningful speech, a trigger instruction, or a provider-neutral
-        // observation completed by the failed attempt. Activity has already retained finalized
-        // filler, so an empty fresh attempt has no value to send and needs no synthetic placeholder.
-        if turnMessages.isEmpty {
-            let preview = delta.lines.isEmpty
-                ? "nothing new"
-                : String(delta.lines.map(\.text).joined(separator: " · ").prefix(80))
-            jlog("… skipped as filler (\(preview)) — not calling the brain")
-            commitTranscript(through: delta.upTo)
-            return AttemptExecution(id: attemptID, result: .skipped(.skippedFillerOnly))
-        }
-        let historyBase: [ChatMessage] = [.system(JarvisPrompts.Coach.system)] + history.snapshot()
-
-        if reason == .manualHint && !work.manualHintPrepared {
-            if let prompt = context.promptLine {
-                jlog("⌨️ hint shortcut — \(prompt)")
-                activity?.record(.manualHint(prompt: prompt))
-            }
-            let screen = self.screen
-            let shot = await Self.captureScreen(using: screen, selecting: attempt.plan.screen)
-            if Task.isCancelled {
-                jlog("… attempt cancelled (stopped) after capture")
-                return AttemptExecution(id: attemptID, result: .cancelled)
-            }
-            if let shot {
-                jlog("👁 looking at your screen")
-                activity?.record(.screenViewed(imageBase64JPEG: shot.imageBase64))
-                var observations: [ChatMessage] = [.userImage(shot.imageBase64)]
-                turnMessages.append(.userImage(shot.imageBase64))
-                if let text = shot.recognizedText {
-                    jlog("🔤 read \(text.count(where: { $0 == "\n" }) + 1) lines of on-screen text")
-                    let observation = ChatMessage.user(JarvisPrompts.Coach.recognizedText(text))
-                    observations.append(observation)
-                    turnMessages.append(observation)
-                }
-                work.observations = observations
-            } else {
-                jlog("👁 screenshot failed")
-                activity?.record(.screenViewFailed)
-                work.observations = [
-                    .user(JarvisPrompts.Coach.manualHintCaptureFailed),
-                ]
-            }
-            work.manualHintPrepared = true
-        }
-
-        let toolChoice: ToolChoice =
-            reason == .manualHint ? .force(speakTool.name) : .required
-        jlog("💭 thinking… [\(attempt.target.provider.displayName)]")
-
-        var requestPhase: CoachingAttemptAuditEvent.RequestPhase = .initial
-        var requestSequence = 1
-        let conversation: any BrainConversation
-        do {
-            let requestContext = CoachingRequestAttribution.context(
-                attemptID: attemptID,
-                wake: work.wake,
-                reason: reason,
-                phase: requestPhase,
-                sequence: requestSequence)
-            conversation = try await CoachingRequestAttribution.$current.withValue(requestContext) {
-                try await attempt.brain.makeConversation()
-            }
-        } catch {
-            if Task.isCancelled || error is CancellationError {
-                jlog("… attempt cancelled (interrupted)")
-                return AttemptExecution(id: attemptID, result: .cancelled)
-            }
-            let failure = BrainFailure(error)
-            jlog("Jarvis coach: brain conversation failed on \(reason) via "
-                 + "\(attempt.target.provider.displayName): \(failure.detail)")
-            return AttemptExecution(
-                id: attemptID,
-                result: .failed(outcome: .brainError, failure: failure, work: work))
-        }
-
-        let result: AttemptResult = await { () async -> AttemptResult in
-            var iterations = 0
-            while iterations < maxToolIterations {
-                iterations += 1
-                let response: BrainResponse
-                do {
-                    let requestContext = CoachingRequestAttribution.context(
-                        attemptID: attemptID,
-                        wake: work.wake,
-                        reason: reason,
-                        phase: requestPhase,
-                        sequence: requestSequence)
-                    response = try await CoachingRequestAttribution.$current.withValue(requestContext) {
-                        try await conversation.respond(
-                            messages: historyBase + turnMessages,
-                            tools: coachTools,
-                            toolChoice: toolChoice)
-                    }
-                } catch {
-                    if Task.isCancelled || error is CancellationError {
-                        jlog("… attempt cancelled (interrupted)")
-                        return .cancelled
-                    }
-                    let failure = BrainFailure(error)
-                    jlog("Jarvis coach: brain request failed on \(reason) via "
-                         + "\(attempt.target.provider.displayName): \(failure.detail)")
-                    return .failed(outcome: .brainError, failure: failure, work: work)
-                }
-
-                if Task.isCancelled {
-                    jlog("… attempt cancelled (stopped) mid-think")
-                    return .cancelled
-                }
-
-                // Incomplete output cannot prove a terminal action, even if it happens to contain one.
-                // Do not render a partial tip or execute a partial tool decision.
-                if let incompleteReason = response.incompleteReason {
-                    jlog("⚠️ response incomplete (\(incompleteReason)) — scheduling fresh attempt")
-                    return .failed(
-                        outcome: .truncated,
-                        failure: BrainFailure(
-                            disposition: .temporary,
-                            detail: "incomplete response: \(incompleteReason)"),
-                        work: work)
-                }
-
-                guard let call = response.toolCalls.first else {
-                    jlog("⚠️ required coaching action missing — scheduling fresh attempt")
-                    return .failed(
-                        outcome: .brainError,
-                        failure: BrainFailure(
-                            disposition: .temporary,
-                            detail: "provider returned no required coaching tool call"),
-                        work: work)
-                }
-
-                switch call {
-                case .captureScreen(let callID):
-                    let screen = self.screen
-                    let shot = await Self.captureScreen(using: screen, selecting: attempt.plan.screen)
-                    if Task.isCancelled {
-                        jlog("… attempt cancelled (stopped) after capture")
-                        return .cancelled
-                    }
-                    if let shot {
-                        jlog("👁 looking at your screen")
-                        activity?.record(.screenViewed(imageBase64JPEG: shot.imageBase64))
-                        if let text = shot.recognizedText {
-                            jlog("🔤 read \(text.count(where: { $0 == "\n" }) + 1) lines of on-screen text")
-                        }
-                        work.observations = [
-                            .user(JarvisPrompts.Coach.captureResult(
-                                recognizedText: shot.recognizedText
-                            )),
-                            .userImage(shot.imageBase64),
-                        ]
-                    } else {
-                        jlog("👁 screenshot failed")
-                        activity?.record(.screenViewFailed)
-                        work.observations = [
-                            .user(JarvisPrompts.Coach.earlierCaptureFailed),
-                        ]
-                    }
-
-                    // Provider-specific linkage remains inside this attempt only.
-                    if !response.outputItemsJSON.isEmpty {
-                        turnMessages.append(.rawItems(response.outputItemsJSON))
-                    } else {
-                        turnMessages.append(.assistantToolCalls(response.rawToolCalls))
-                    }
-                    if let shot {
-                        turnMessages.append(.init(
-                            role: .tool,
-                            text: JarvisPrompts.Coach.captureResult(
-                                recognizedText: shot.recognizedText
-                            ),
-                            toolCallId: callID))
-                        turnMessages.append(.userImage(shot.imageBase64))
-                    } else {
-                        turnMessages.append(.init(
-                            role: .tool,
-                            text: JarvisPrompts.Coach.captureFailed,
-                            toolCallId: callID))
-                    }
-                    requestPhase = .captureScreenContinuation
-                    requestSequence += 1
-
-                case .speak(let callID, let lines):
-                    if Task.isCancelled {
-                        jlog("… attempt cancelled (stopped) before speaking")
-                        return .cancelled
-                    }
-                    jlog("💬 \(lines.joined(separator: " "))")
-                    activity?.record(.tip(lines: lines))
-                    overlay.render(
-                        lines,
-                        perLineSeconds: lines.map {
-                            OverlayTiming.displaySeconds(for: $0, config: config)
-                        })
-                    turnMessages.append(.assistantToolCalls(response.rawToolCalls))
-                    turnMessages.append(.init(
-                        role: .tool,
-                        text: JarvisPrompts.Coach.tipShown,
-                        toolCallId: callID))
-                    history.commit(turnMessages)
-                    commitTranscript(through: delta.upTo)
-                    return .completed(.spoke)
-
-                case .staySilent:
-                    if Task.isCancelled {
-                        jlog("… attempt cancelled (stopped) before recording silence")
-                        return .cancelled
-                    }
-                    jlog("… nothing useful to add, staying silent")
-                    activity?.record(.stayedSilent)
-                    commitIfWorthKeeping(turnMessages, deltaText: substantiveDeltaText)
-                    commitTranscript(through: delta.upTo)
-                    return .completed(.silentByModel)
-                }
-            }
-
-            jlog("⚠️ tool loop exhausted — scheduling fresh attempt")
-            return .failed(
-                outcome: .exhausted,
-                failure: BrainFailure(
-                    disposition: .temporary,
-                    detail: "coaching tool loop exhausted"),
-                work: work)
-        }()
-
-        await conversation.finish()
-        if case .completed = result {
-            startCompactionIfIdle(using: attempt.summarizer ?? attempt.brain)
-        }
-        return AttemptExecution(id: attemptID, result: result)
-    }
-
-    private func commitIfWorthKeeping(_ turn: [ChatMessage], deltaText: String) {
-        guard !deltaText.isEmpty || turn.count > 1 else { return }
-        history.commit(turn)
-    }
-
-    /// Screen capture is an OS-bound synchronous edge, so run it off the cooperative executor.
-    /// Cancellation asks the capture adapter to terminate its helper, then waits for `capture()` to
-    /// return so the attempt cannot outlive cleanup of screen-derived files.
-    private static func captureScreen(
-        using screen: ScreenCapturing,
-        selecting selection: ScreenCaptureSelection
-    ) async -> ScreenSnapshot? {
-        let capture = Task.detached(priority: .userInitiated) { () -> ScreenSnapshot? in
-            guard !Task.isCancelled else { return nil }
-            return screen.capture(selection)
-        }
-        return await withTaskCancellationHandler {
-            // Await the producer itself. A cancelled AsyncStream consumer returns immediately and
-            // would let conversation/session drain race the helper's exit and JPEG cleanup.
-            await capture.value
-        } onCancel: {
-            capture.cancel()
-            screen.cancelCapture()
-        }
-    }
-
-    // MARK: - History compaction
-
-    /// Compaction is auxiliary, so it runs off the attempt path: awaiting it here would spend its
-    /// whole budget as dead air, batching newly finalized speech behind a summary nobody is waiting
-    /// for. One run at a time — two overlapping runs each replace `0..<prefixCount`, so the second
-    /// would destroy the first's summary along with real turns. A skipped or failed run simply
-    /// retries from a fresh prefix after the next completed attempt.
-    private func startCompactionIfIdle(using client: BrainClient) {
-        // Check the threshold before spawning: most completed attempts are nowhere near it, and a
-        // task per attempt would be pure churn on the shared executor.
-        guard history.estimatedTokens > config.historyCompactionTokenThreshold else { return }
-        stateLock.lock()
-        // Stop is final. A turn suspended in `conversation.finish()` when teardown ran resumes after
-        // `cancelBackgroundWork` already looked for a task, and would otherwise start a fresh
-        // provider request that nothing is left to drain.
-        guard !backgroundWorkStopped else {
-            stateLock.unlock()
-            return
-        }
-        guard !isCompacting else {
-            // Coalesce rather than drop. An attempt landing mid-summary is also the attempt whose
-            // fresh OCR can invalidate that summary, so without this the rejected pass and the
-            // skipped request cancel each other out and history stays over the threshold.
-            compactionRequested = true
-            stateLock.unlock()
-            return
-        }
-        isCompacting = true
-        compactionRequested = false
-        let task = Task.detached(priority: .utility) { [self] in
-            await runCompactionPasses(using: client)
-        }
-        compactionTask = task
-        stateLock.unlock()
-    }
-
-    /// Run compaction until nothing more is pending, so a pass that ends without compacting — a
-    /// summary rejected as stale, or one skipped while another ran — is retried instead of waiting
-    /// for a further completed attempt.
-    private func runCompactionPasses(using client: BrainClient) async {
-        while true {
-            await compactIfNeeded(using: client)
-            let overThreshold = history.estimatedTokens > config.historyCompactionTokenThreshold
-            let runAgain: Bool = stateLock.withLock {
-                guard compactionRequested, overThreshold, !Task.isCancelled else {
-                    isCompacting = false
-                    compactionTask = nil
-                    return false
-                }
-                compactionRequested = false
-                return true
-            }
-            if !runAgain { return }
-        }
-    }
-
-    /// Cancel background work that must not outlive the session, handing back the task so teardown
-    /// can drain it before sealing the audit.
-    ///
-    /// Compaction runs off the attempt path, so `TurnTaskBox` does not own it. Without this a
-    /// summary keeps a provider process alive and billing after Stop, and can still be writing when
-    /// the session audit closes.
-    @discardableResult
-    public func cancelBackgroundWork() -> Task<Void, Never>? {
-        stateLock.lock()
-        backgroundWorkStopped = true
-        compactionRequested = false
-        let task = compactionTask
-        compactionTask = nil
-        stateLock.unlock()
-        task?.cancel()
-        return task
-    }
-
-    /// Auxiliary compaction fails soft and never counts against provider route health.
-    private func compactIfNeeded(using client: BrainClient) async {
-        guard history.estimatedTokens > config.historyCompactionTokenThreshold else { return }
-        guard let (oldest, count, revision) = history.compactionPrefix() else { return }
-        do {
-            let response = try await client.respond(
-                messages: [
-                    .system(JarvisPrompts.HistorySummary.system),
-                    .user(JarvisPrompts.HistorySummary.input(oldest)),
-                ],
-                tools: [],
-                toolChoice: .auto)
-            guard let summary = response.outputText, !summary.isEmpty else {
-                jlog("… memory compaction returned nothing — keeping full history for now")
-                return
-            }
-            // A summary that lands as teardown cancels must not mutate history on the way out: the
-            // session is over and the provider may simply have won the race.
-            guard !Task.isCancelled else { return }
-            guard history.compact(prefixCount: count, summary: summary, revision: revision) else {
-                jlog("… discarded a summary written against superseded screen text — will retry later")
-                return
-            }
-            jlog("… condensed session memory to ~\(history.estimatedTokens) tokens")
-        } catch {
-            jlog("… memory compaction failed (will retry later): \(error.localizedDescription)")
-        }
     }
 
 }
